requirements.txt
# API and Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
aiohttp>=3.8.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml==4.9.3

# Data Processing and Analysis
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0
joblib==1.3.2
matplotlib>=3.5.0
seaborn>=0.11.0

# Environment and Configuration
python-dotenv>=0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Machine learning
xgboost>=2.0.0
tensorflow>=2.15.0

# Web framework
flask>=3.0.0

# Data visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0

webdriver-manager==4.0.2 